[Throwable] Update getTraceCode method.
